Add raw filter, fix controllers names, override SelectAll
