+++ src/Controller/SceneController.php
@@ -5,10 +5,6 @@ namespace App\Controller;
 use App\Model\SceneManager;
 use App\Model\DialogueManager;
 use App\Model\ChoiceManager;
-use App\Model\CharacterManager;
-use App\Model\StoryManager;
-
-// use App\Model\UserSaveManager; à venir
 
 class SceneController extends AbstractController
 {
@@ -16,12 +12,6 @@ class SceneController extends AbstractController
     private $dialogueManager;
 
     private $choiceManager;
-
-    private $characterManager;
-
-    private $storyManager;
-
-    private const TARGET_DIR = 'assets/images/backgrounds/';
 
     public const EXTENSIONS_ALLOWED = ['jpg', 'jpeg', 'png', 'webp', 'svg'];
     public const MAX_UPLOAD_SIZE = 5000000;
@@ -34,118 +24,8 @@ class SceneController extends AbstractController
         $this->sceneManager = new SceneManager();
         $this->dialogueManager = new DialogueManager();
         $this->choiceManager = new ChoiceManager();
-        $this->characterManager = new CharacterManager();
-        $this->storyManager = new StoryManager();
-    }
-    public function add(int $storyId): ?string
-    {
-        $errors = [];
-
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $scene = array_map('htmlentities', array_map('trim', $_POST));
-            $scene['story_id'] = $storyId;
-
-            if (!empty($_FILES['background']['full_path'])) {
-                $uploadErrors = $this->handleBackgroundUpload();
-                $errors = array_merge($errors, $uploadErrors);
-
-                if (empty($uploadErrors)) {
-                    $scene['background'] = basename($_FILES['background']['name']);
-                }
-            } else {
-                $scene["background"] = "";
-            }
-
-            $validationErrors = $this->validateScene($scene);
-            $errors = array_merge($errors, $validationErrors);
-
-            if (empty($errors)) {
-                $scene['background'] = basename($_FILES['background']['name']);
-                $id = $this->sceneManager->insert($scene);
-            } else {
-                error_log('Erreurs lors de l\'ajout de la scène: ' . implode(', ', $errors));
-                // Insère la scène sans background en cas d'erreur
-                $id = $this->sceneManager->insert($scene);
-            }
-
-            header('Location:/story/engine/scene/show?' . http_build_query(['story_id' => $storyId, 'id' => $id]));
-            return null;
-        }
-
-        return $this->twig->render('SceneCreation/add.html.twig');
     }
 
-    public function update(): ?string
-    {
-        $scene = [];
-        $errors = [];
-
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $scene = array_map('htmlentities', array_map('trim', $_POST));
-            $previousSettings = $this->sceneManager->selectOneById($scene['scene_id']);
-
-            if (!empty($_FILES['background']['full_path'])) {
-                $uploadErrors = $this->handleBackgroundUpload();
-                $errors = array_merge($errors, $uploadErrors);
-                if (empty($uploadErrors)) {
-                    $scene['background'] = basename($_FILES['background']['name']);
-                } else {
-                    $scene['background'] = $previousSettings['background'];
-                }
-            } else {
-                $scene['background'] = $previousSettings['background'];
-            }
-
-            if (empty($scene['name'])) {
-                $scene['name'] = $previousSettings['name'];
-            }
-
-            $validationErrors = $this->validateScene($scene);
-            $errors = array_merge($errors, $validationErrors);
-
-
-            if (empty($errors)) {
-                $this->sceneManager->update($scene);
-            } else {
-                error_log('Erreur lors de l\'edition de la scene : ' . implode(', ', $errors));
-            }
-
-            header('Location:/story/engine/scene/show?story_id='
-            . $scene['story_id'] . '&id=' . $scene['scene_id']);
-        }
-        return null;
-    }
-
-
-    public function showCreation(string $storyId, string $id): string
-    {
-        $scene = $this->sceneManager->selectOneById((int) $id);
-        $story = $this->storyManager->selectOneById((int) $storyId);
-        $dialogues = $this->dialogueManager->selectAllByScene((int) $id);
-        $choices = $this->choiceManager->selectAllByScene((int) $id);
-        $characters = $this->characterManager->selectByStory($storyId);
-        $allscenes = $this->sceneManager->selectAllByStory($storyId);
-
-        return $this->twig->render(
-            'SceneCreation/show.html.twig',
-            [
-                'scene' => $scene,
-                'story' => $story,
-                'dialogues' => $dialogues,
-                'choices' => $choices,
-                'characters' => $characters,
-                'allscenes' => $allscenes,
-            ]
-        );
-    }
-
-    public function delete(string $storyId, string $sceneId): ?string
-    {
-        $this->sceneManager->delete((int) $sceneId);
-
-        header('Location:/story/engine/show?id=' . $storyId);
-        return null;
-    }
 
     public function show(int $sceneId): string
     {
@@ -157,9 +37,9 @@ class SceneController extends AbstractController
             exit();
         }
 
-        $dialogues = $this->dialogueManager->selectAllByScene($sceneId);
+        $dialogues = $this->dialogueManager->selectAll($sceneId);
 
-        $choices = $this->choiceManager->selectAllByScene($sceneId);
+        $choices = $this->choiceManager->selectAll($sceneId);
 
         return $this->twig->render('Scene/show.html.twig', [
             'scene' => $scene,
@@ -167,8 +47,6 @@ class SceneController extends AbstractController
             'choices' => $choices
         ]);
     }
-
-
 
     public function showFirstScene(int $storyId): string
     {
@@ -181,47 +59,5 @@ class SceneController extends AbstractController
         }
 
         return $this->show((int) $sceneId['scene_id']);
-    }
-
-    public function handleBackgroundUpload(): array
-    {
-        $errors = [];
-
-        $targetFile = self::TARGET_DIR . basename($_FILES['background']['name']);
-        $typeFile = strtolower(pathinfo($targetFile, PATHINFO_EXTENSION));
-        $dimensionsImage = getimagesize($_FILES['background']['tmp_name']);
-
-
-        if (!$dimensionsImage) {
-            $errors[] = 'Votre background n\'est pas une image';
-        }
-
-        if ($_FILES['background']['size'] > self::MAX_UPLOAD_SIZE) {
-            $errors[] = 'Votre background ne peut pas dépasser ' . self::MAX_UPLOAD_SIZE / 1000000 . 'Mo';
-        }
-
-        if (!in_array($typeFile, self::EXTENSIONS_ALLOWED)) {
-            $errors[] = 'Votre background n\'as pas le bon format ('
-                . implode(', ', self::EXTENSIONS_ALLOWED) . ')';
-        }
-
-        if (!move_uploaded_file($_FILES['background']['tmp_name'], $targetFile)) {
-            $errors[] = 'Erreur lors du déplacement du fichier de background';
-            error_log('Erreur lors du déplacement du fichier de background: ' . $_FILES['background']['error']);
-        }
-
-        return $errors;
-    }
-
-    public function validateScene(array $scene): array
-    {
-        $errors = [];
-
-        if (strlen($scene["name"]) >= self::MAX_SCENE_TITLE_LENGTH) {
-            $errors[] = "Le titre de votre scene est trop long, maximum : "
-             . self::MAX_SCENE_TITLE_LENGTH . " caractères.";
-        }
-
-        return $errors;
     }
 }
