+++ src/View/SceneCreation/_character_form.html.twig
@@ -10,10 +10,10 @@
       <button class="edit_character">Editer</button>
       <form method="post" enctype="multipart/form-data" action="/story/engine/character/update" class="edit_character_form">
         <img src="/assets/images/sprites/{{character.sprite}}">
-        <input type="hidden" name="story_id" value={{ story.id }}>
-        <input type="hidden" name="scene_id" value={{ scene.id }}>
-        <input type="hidden" name="character_id" value={{ character.id }}>
-        <input type="text" name="character_name" id="character_name" value={{character.name}}>
+        <input type="hidden" name="story_id" value={{ story.id}}>
+        <input type="hidden" name="scene_id" value={{ scene.id}}>
+        <input type="hidden" name="character_id" value={{ character.id}}>
+        <input type="text" name="character_name" id="character_name" value={{character.name | raw}}>
         <label for="sprite">sprite</label>
         <input type="file" name="sprite" id="sprite">
         <input type="submit">
