+++ src/View/SceneCreation/_choice_form.html.twig
@@ -1,12 +1,12 @@
 <ul>
   {% for choice in choices %}
   <li>
-{{ choice.body }}
+{{ choice.body | raw }}
   {% if choice.next_scene_id == 0 %}
-    {{ choice.next_scene_name }}
+    {{ choice.next_scene_name | raw }}
   {% else %}
-    <a href="/story/editor/scene/show?story_id={{ story.id }}&id={{ choice.next_scene_id }}">
-    {{ choice.next_scene_name }}</a>
+    <a href="/story/engine/scene/show?story_id={{ story.id }}&id={{ choice.next_scene_id }}">
+    {{ choice.next_scene_name | raw }}</a>
   {% endif %}
     <button onclick="if(confirm('Êtes vous sûr de vouloir supprimer ce choix ?')) { window.location.href='/story/engine/choice/delete?story_id={{story.id}}&scene_id={{scene.id}}&id={{choice.id}}';}"> &#x1F5D1;</button>
     <button class="edit_choice">Editer</button>
@@ -14,11 +14,11 @@
     <input type="hidden" name="id" value="{{ choice.id }}">
     <input type="hidden" name="scene_id" value="{{ scene.id }}">
     <input type="hidden" name="story_id" value="{{ story.id }}">
-    <input type="text" name="choice_body" id="choice_body" value="{{ choice.body }}">
+    <input type="text" name="choice_body" id="choice_body" value="{{ choice.body | raw }}">
     <select name="next_scene" id="next_scene">
         {% for onescene in allscenes %}
         <option value="{{ onescene.id }}" {% if onescene.id == choice.next_scene_id %}selected{% endif %}>
-        {{ onescene.name }}
+        {{ onescene.name | raw }}
         </option>
         {% endfor %}
     </select>
@@ -37,7 +37,7 @@
       <option value="0">-- À définir --</option>
       {% for onescene in allscenes %}
       <option value={{ onescene.id }}>
-        {{ onescene.name }}
+        {{ onescene.name | raw }}
       </option>
       {% endfor %}
     </select>
