+++ src/View/SceneCreation/_dialogue_form.html.twig
@@ -1,10 +1,10 @@
 <ul>
     {% for dialogue in dialogues %}
     <li>
-      {{ dialogue.name }} :
-      {{ dialogue.body }}
+      {{ dialogue.name | raw }} :
+      {{ dialogue.body | raw }}
         <button
-        onclick="if(confirm('Êtes vous sûr de vouloir supprimer ce choix ?'))
+        onclick="if(confirm('Êtes vous sûr de vouloir supprimer ce dialogue ?'))
         { window.location.href='/story/engine/dialogue/delete?story_id={{story.id}}&scene_id={{scene.id}}&id={{dialogue.id}}';}">
         &#x1F5D1;
         </button>
@@ -13,11 +13,11 @@
             <input type="hidden" name="story_id" value={{ story.id }}>
             <input type="hidden" name="scene_id" value={{ scene.id }}>
             <input type="hidden" name="dialogue_id" value={{ dialogue.id }}>
-            <input type="textarea" name="dialogue_body" id="dialogue_body" value="{{ dialogue.body }}">
+            <input type="textarea" name="dialogue_body" id="dialogue_body" value="{{ dialogue.body | raw }}">
             <select name="character_id" id="character_id">
             {% for character in characters %}
                 <option value={{ character.id }} {% if onescene.id == choice.next_scene_id %}selected{% endif %} >
-                {{ character.name }}
+                {{ character.name | raw }}
                 </option>
             {% endfor %}
             </select>
@@ -36,7 +36,7 @@
             <label for = "character_id"></label>
             <select name = "character_id" id = "character_id">
             {% for character in characters %}
-                <option  value = "{{ character.id }}"> {{ character.name }} </option>
+                <option  value = "{{ character.id }}"> {{ character.name | raw }} </option>
             {% endfor %}
         </select>
         <input type ="submit">
