+++ src/View/SceneCreation/show.html.twig
@@ -9,13 +9,13 @@
 <main id="mainScenes">
   <section>
   <h1>Histoire :
-    {{ story.name }}
+    {{ story.name | raw }}
   </h1>
   <a href="/story/engine/show?id={{ story.id }}">Retour à l'histoire</a>
-  <img height=300 width=300 src = "/assets/images/backgrounds/{{scene.background}}" alt="Background">
+  <img height=300 width=300 src = "/assets/images/backgrounds/{{scene.background | raw}}" alt="Background">
     {% include 'SceneCreation/_scene_edit_form.html.twig' %}
 
-    <h1>{{scene.name}}</h1>
+    <h1>{{scene.name | raw }}</h1>
 
     <h2>Personnages :</h2>
     {% include 'SceneCreation/_character_form.html.twig' %}
