+++ src/View/StoryCreation/index.html.twig
@@ -15,7 +15,7 @@
             <article class="storiesDetails">
                 <div class="storyDetails">
                     <div class="name">
-                        <h2><a href="/story/scene/start?storyId={{ story.id }}">{{ story.name }}</a></h2>
+                        <h2><a href="/story/scene/start?storyId={{ story.id }}">{{ story.name | raw }}</a></h2>
                     </div>
                     <div class="details">
                         <h3>Nb scenes : </h3>
@@ -30,7 +30,7 @@
                 </div>
                 <div class="storyDelete">
                     <a
-                        onclick="if(confirm('Êtes vous sûr de vouloir supprimer ce choix ?'))
+                        onclick="if(confirm('Êtes vous sûr de vouloir supprimer cette histoire ?'))
                         { window.location.href='/story/engine/delete?id={{story.id}}';}">
                         <img src="/assets/images/icones/trash.svg" alt="trash icon" class="icon">
                     </a>
