+++ src/View/StoryCreation/show.html.twig
@@ -12,7 +12,7 @@
             <img src="/assets/images/icones/home.svg" alt="plow icon"> Home
         </a>
     </div>
-    <h1>{{story.name}}</h1>
+    <h1>{{story.name | raw}}</h1>
     <nav id="nav">
         <a href="/story/engine"><h2 id="navScene">Scènes</h2></a>
         <a href="/story/engine/character/show"><h2 id="navCharacter">Personnages</h2></a>
@@ -22,7 +22,7 @@
             <article class="sceneList">
                 <div class="name">
                     <a href="/story/scene/show?scene_id={{ scene.id }}">
-                        {{ scene.name }}
+                        {{ scene.name | raw }}
                     </a>
                 </div>
                 <div id="sceneEdit">
@@ -51,7 +51,7 @@
             <article class="ressourceList">
                 <div class="name">
                     <a href="/story/scene/show?story_id={{ story.id }}&id={{ scene.id }}">
-                        {{ ressource.name }}
+                        {{ ressource.name | raw }}
                     </a>
                 </div>
                 <div id="ressourceEdit">
