+++ src/View/layout.html.twig
@@ -1,7 +1,6 @@
 <!doctype html>
 <html lang="fr">
 
-
 <head>
   <meta charset="utf-8">
   <meta http-equiv="x-ua-compatible" content="ie=edge">
