+++ src/routes.php
@@ -12,16 +12,16 @@ return [
     'story' => ['StoryController', 'index'],
     'story/scene/start' => ['SceneController', 'showFirstScene', ['storyId']],
     'story/scene/show' => ['SceneController', 'show', ['scene_id']],
-    'story/engine' => ['StoryController', 'indexCreation'],
     // Création d'une histoire
-    'story/engine/add' => ['StoryController', 'add',],
-    'story/engine/show' => ['StoryController', 'showCreation', ['id']],
-    'story/engine/delete' => ['StoryController', 'delete', ['id']],
+    'story/engine' => ['StoryCreationController', 'index'],
+    'story/engine/add' => ['StoryCreationController', 'add',],
+    'story/engine/show' => ['StoryCreationController', 'show', ['id']],
+    'story/engine/delete' => ['StoryCreationController', 'delete', ['id']],
     // Création d'une scène
-    'story/engine/scene/add' => ['SceneController', 'add', ['story_id']],
-    'story/engine/scene/show' => ['SceneController', 'showCreation', ['story_id', 'id']],
-    'story/engine/scene/delete' => ['SceneController', 'delete', ['story_id', 'id']],
-    'story/engine/scene/update' => ['SceneController', 'update', ['story_id', 'id']],
+    'story/engine/scene/add' => ['SceneCreationController', 'add', ['story_id']],
+    'story/engine/scene/show' => ['SceneCreationController', 'show', ['story_id', 'id']],
+    'story/engine/scene/delete' => ['SceneCreationController', 'delete', ['story_id', 'id']],
+    'story/engine/scene/update' => ['SceneCreationController', 'update', ['story_id', 'id']],
     // Création d'un personnage
     'story/engine/character/show' => ['CharacterController', 'show', ['story_id']],
     'story/engine/character/add' => ['CharacterController', 'add'],
